% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%*\%.mmatrix}
\alias{\%*\%.mmatrix}
\title{Matrix product operator for mmatrix class}
\usage{
\method{\%*\%}{mmatrix}(lhs, rhs)
}
\description{
Matrix product operator for mmatrix class
}
\author{
Zach Vig
}
