% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_predict.R
\name{mm_predict}
\alias{mm_predict}
\title{Schema-safe transform for new data}
\usage{
mm_predict(spec, newdata, unknown = c("other", "zero", "error"))
}
\arguments{
\item{spec}{spec from mmatrix()}

\item{newdata}{data.frame}

\item{unknown}{c("other","zero","error")}
}
\value{
sparseMatrix with the same columns/order as training
}
\description{
Schema-safe transform for new data
}
