% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-mmatrix_spec.R
\name{print.mmatrix_spec}
\alias{print.mmatrix_spec}
\title{Pretty printing for mmatrix_spec objects}
\usage{
\method{print}{mmatrix_spec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mmatrix_spec}.}

\item{...}{Ignored (for compatibility with generic).}
}
\description{
Pretty printing for mmatrix_spec objects
}
