% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mmatrix_spec.R
\name{summary.mmatrix_spec}
\alias{summary.mmatrix_spec}
\title{Summary for mmatrix_spec objects}
\usage{
\method{summary}{mmatrix_spec}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mmatrix_spec}.}

\item{...}{Ignored.}
}
\value{
A list with basic summary info (invisible).
}
\description{
Summary for mmatrix_spec objects
}
