% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Create an mmatrix, a base R matrix with simplied printing}
\usage{
m(...)
}
\arguments{
\item{...}{The matrix elements separated by commas. Separate new rows using the %;% operator.}
}
\value{
A matrix with mmatrix printing properties
}
\description{
Create an mmatrix, a base R matrix with simplied printing
}
\examples{
\code{m(1, 2, 3 \%;\% 4, 5, 6)}

}
\author{
Zach Vig
}
